+++ package.json
@@ -10,9 +10,11 @@
   },
   "dependencies": {
     "bootstrap": "^5.3.3",
+    "dompurify": "^3.1.0",
     "gsap": "^3.12.5",
     "microcms-js-sdk": "^3.0.0",
     "next": "14.1.2",
+    "node-fetch": "^3.3.2",
     "nodemailer": "^6.9.13",
     "react": "^18",
     "react-bootstrap": "^2.10.2",
@@ -21,12 +23,16 @@
     "sass": "^1.71.1"
   },
   "devDependencies": {
+    "@types/dompurify": "^3.0.5",
+    "@types/jsdom": "^21.1.6",
     "@types/node": "^20",
     "@types/nodemailer": "^6.4.14",
     "@types/react": "^18",
     "@types/react-dom": "^18",
+    "@types/react-google-recaptcha": "^2.1.9",
     "eslint": "^8",
     "eslint-config-next": "14.1.2",
+    "react-google-recaptcha": "^3.1.0",
     "typescript": "^5"
   }
 }
