+++ src/pages/member/index.tsx
@@ -2,6 +2,8 @@ import React from "react";
 import type { GetStaticProps, InferGetStaticPropsType, NextPage } from "next";
 import { client } from "../../../libs/client";
 import MemberList from "@/components/sections/member/MemberList";
+import styles from "@/styles/layout/TopPage.module.scss";
+import Hero from "@/components/hero/Hero";
 
 type Props = InferGetStaticPropsType<typeof getStaticProps>;
 
@@ -16,7 +18,8 @@ export const getStaticProps: GetStaticProps = async () => {
 
 const Member: NextPage<Props> = ({ member }) => {
   return (
-    <div>
+    <div className={styles.container}>
+      <Hero />
       <MemberList members={member} />
     </div>
   );
