+++ src/styles/sections/member/MemberList.module.scss
@@ -1,41 +1,42 @@
 @use "../../globals.scss" as *;
 
 .section {
+  height: 100%;
   margin: 0;
-  padding: 0;
+  display: flex;
+  align-items: center;
+  justify-content: center;
   font-family: $noto;
 }
 .container {
-  position: relative;
-  background-color: $wh;
-  width: 1080px;
-  margin: 40px auto 0;
-  padding: 80px;
-  border-radius: 20px;
+  text-align: center;
+  padding: 0;
 }
 
 .card {
   display: flex;
-  flex-wrap: wrap;
-  gap: 50px;
-  justify-content: inline-start;
+  align-items: center;
+  justify-content: center;
+  background-color: $wh;
+  box-shadow: 0 0 8px rgba(0, 0, 0, 0.16);
+  border-radius: 20px;
+  margin: 20px;
+  gap: 20px;
 }
 .link {
-  display: flex;
-  align-items: flex-start;
-  gap: 40px;
-  padding: 24px 0;
+  margin: 20px;
 }
 
 .image {
-  width: 380px;
-  height: 220px;
+  width: 400px;
+  height: 200px;
+
   border-radius: 20px;
-  box-shadow: 5px 5px 20px $bk;
+  box-shadow: 0 0 8px rgba(0, 0, 0, 0.16);
   object-fit: cover;
 }
 .content {
-  margin-top: 20px;
+  margin-top: 10px;
 }
 
 .title {
@@ -55,8 +56,7 @@
 
 .past {
   max-width: 380px;
-  margin: 50px;
-  margin-left: calc(50% - 190px);
+  margin: 50px auto;
   font-size: 2rem;
   font-weight: $bold;
   color: $wh;
@@ -64,6 +64,32 @@
   border-radius: 20px;
   display: flex;
   justify-content: center;
+  align-items: center;
   padding: 20px;
   border: solid 0.2rem $bk;
+}
+
+@media screen and (max-width: 834px) {
+  .container {
+    width: 680px;
+  }
+  .card {
+    gap: 10px;
+  }
+  .image {
+    width: 230px;
+    height: 120px;
+  }
+}
+@media screen and (max-width: 430px) {
+  .container {
+    width: 330px;
+  }
+  .card {
+    gap: 0px;
+  }
+  .image {
+    width: 100px;
+    height: 80px;
+  }
 }
